requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2026.1.4
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.9
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.14.0
packaging==25.0
psycopg2-binary==2.9.11
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pyttsx3==2.99
sniffio==1.3.1
SpeechRecognition==3.14.5
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.11.0
